Move helper methods into TestCase class
