+++ tests/Content/TestCase.php
@@ -3,6 +3,7 @@
 namespace Kirby\Content;
 
 use Kirby\Cms\App;
+use Kirby\Data\Data;
 use Kirby\Filesystem\Dir;
 use Kirby\TestCase as BaseTestCase;
 
@@ -11,6 +12,66 @@ class TestCase extends BaseTestCase
 	public const TMP = KIRBY_TMP_DIR;
 
 	protected $model;
+
+	public function assertContentFileExists(string|null $language = null, VersionId|null $versionId = null)
+	{
+		$this->assertFileExists($this->contentFile($language, $versionId));
+	}
+
+	public function assertContentFileDoesNotExist(string|null $language = null, VersionId|null $versionId = null)
+	{
+		$this->assertFileDoesNotExist($this->contentFile($language, $versionId));
+	}
+
+	public function contentFile(string|null $language = null, VersionId|null $versionId = null): string
+	{
+		return
+			$this->model->root() .
+			// add the changes folder
+			($versionId?->value() === 'changes' ? '/_changes/' : '/') .
+			// template
+			'article' .
+			// language code
+			($language === null ? '' : '.' . $language) .
+			'.txt';
+	}
+
+	public function createContentMultiLanguage(): array
+	{
+		Data::write($fileEN = $this->contentFile('en'), $contentEN = [
+			'title'    => 'Title English',
+			'subtitle' => 'Subtitle English'
+		]);
+
+		Data::write($fileDE = $this->contentFile('de'), $contentDE = [
+			'title'    => 'Title Deutsch',
+			'subtitle' => 'Subtitle Deutsch'
+		]);
+
+		return [
+			'en' => [
+				'content' => $contentEN,
+				'file'    => $fileEN,
+			],
+			'de' => [
+				'content' => $contentDE,
+				'file'    => $fileDE,
+			]
+		];
+	}
+
+	public function createContentSingleLanguage(): array
+	{
+		Data::write($file = $this->contentFile(), $content = [
+			'title'    => 'Title',
+			'subtitle' => 'Subtitle'
+		]);
+
+		return [
+			'content' => $content,
+			'file'    => $file
+		];
+	}
 
 	public function setUp(): void
 	{
