+++ tests/Content/VersionTest.php
@@ -13,66 +13,6 @@ class VersionTest extends TestCase
 {
 	public const TMP = KIRBY_TMP_DIR . '/Content.Version';
 
-	public function assertContentFileExists(string|null $language = null, VersionId|null $versionId = null)
-	{
-		$this->assertFileExists($this->contentFile($language, $versionId));
-	}
-
-	public function assertContentFileDoesNotExist(string|null $language = null, VersionId|null $versionId = null)
-	{
-		$this->assertFileDoesNotExist($this->contentFile($language, $versionId));
-	}
-
-	public function contentFile(string|null $language = null, VersionId|null $versionId = null): string
-	{
-		return
-			$this->model->root() .
-			// add the changes folder
-			($versionId?->value() === 'changes' ? '/_changes/' : '/') .
-			// template
-			'article' .
-			// language code
-			($language === null ? '' : '.' . $language) .
-			'.txt';
-	}
-
-	public function createContentMultiLanguage(): array
-	{
-		Data::write($fileEN = $this->contentFile('en'), $contentEN = [
-			'title'    => 'Title English',
-			'subtitle' => 'Subtitle English'
-		]);
-
-		Data::write($fileDE = $this->contentFile('de'), $contentDE = [
-			'title'    => 'Title Deutsch',
-			'subtitle' => 'Subtitle Deutsch'
-		]);
-
-		return [
-			'en' => [
-				'content' => $contentEN,
-				'file'    => $fileEN,
-			],
-			'de' => [
-				'content' => $contentDE,
-				'file'    => $fileDE,
-			]
-		];
-	}
-
-	public function createContentSingleLanguage(): array
-	{
-		Data::write($file = $this->contentFile(), $content = [
-			'title'    => 'Title',
-			'subtitle' => 'Subtitle'
-		]);
-
-		return [
-			'content' => $content,
-			'file'    => $file
-		];
-	}
-
 	/**
 	 * @covers ::content
 	 */
